FindGZIP.cmake working!
I appended a variable `POSIX_ROOT` to the beginning of the search paths, and
defined it, when WIN32 is set, as "C:/Program\ Files/Git", else "".
As a result, it should continue working normally on Linux, and should now working
in Windows environments where Git is installed in the normal directory (maybe
these paths should be hard coded into FindGZIP instead of being hard-coded? we
shall see what works best later; I wanted to make this flexable)