Created platform lib to facilitate portability
Created an `STATIC` library named "platform" with the libraries added
as `INTERFACE`. I'm new to CMake, so hopefully this is the right way
to go! Linked the library in the CMakeLists.txt file in the root
directory to the executable target `pdip_exe`. I also defined a macro
`PRISIZE` which expands to the format specifier for size_t dependant
on what platform pdip is to be compiled for. More to come!