Replaced GNU-specific `%m` format specifier
The `%m` format specifier is a GNU extension that does not require
any additional arguments to be passed and is replaced with the
result of `strerror(errno)`.
Such a thing IS NOT portable, so I replaced every instance
(at least in pdip.c) with `%s` and a corresponding call to
`strerror(errno)`.